# src/specs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ROLE": {
				"type": "string",
				"enum": [
					"user",
					"admin"
				],
				"nullable": false
			},
			"IUserBaseModel": {
				"properties": {
					"id": {},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					},
					"role": {
						"$ref": "#/components/schemas/ROLE"
					},
					"isEmailVerified": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"role",
					"isEmailVerified"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IApiErrorDetail": {
				"properties": {
					"message": {
						"type": "string"
					},
					"field": {
						"type": "string"
					},
					"value": {},
					"path": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IApiError": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/IApiErrorDetail"
						},
						"type": "array"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUserBaseModel.Exclude_keyofIUserBaseModel.id-or-isEmailVerified__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin"
						]
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"role"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewCreatedUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin"
						]
					}
				},
				"required": [
					"name",
					"email",
					"password",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QueryResult_IUserBaseModel_": {
				"properties": {
					"results": {
						"items": {
							"$ref": "#/components/schemas/IUserBaseModel"
						},
						"type": "array"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"totalResults": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"results",
					"page",
					"limit",
					"totalPages",
					"totalResults"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserBodyByAdmin": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"TokenPayload": {
				"properties": {
					"token": {
						"type": "string"
					},
					"expires": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"token",
					"expires"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccessAndRefreshTokens": {
				"properties": {
					"access": {
						"$ref": "#/components/schemas/TokenPayload"
					},
					"refresh": {
						"$ref": "#/components/schemas/TokenPayload"
					}
				},
				"required": [
					"access",
					"refresh"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserWithTokens": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/IUserBaseModel"
					},
					"tokens": {
						"$ref": "#/components/schemas/AccessAndRefreshTokens"
					}
				},
				"required": [
					"user",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_IUserBaseModel.Exclude_keyofIUserBaseModel.id-or-role-or-isEmailVerified__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"NewRegisteredUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string",
						"minLength": 8
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "@kazize/create-ts-starter-1",
		"version": "2.2.0",
		"description": "TypeScript Express.js backend boilerplate on AWS Lambda",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Kazi Ehsan Aziz ",
			"email": "kazizehsan@gmail.com"
		}
	},
	"paths": {
		"/v1/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserBaseModel"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable entity.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"description": "Create users by admin.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewCreatedUser"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueryResult_IUserBaseModel_"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "projectBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserBaseModel"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/v1/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserBaseModel"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Success.",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUserBaseModel"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable entity.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserBodyByAdmin"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"204": {
						"description": "Success."
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithTokens"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"422": {
						"description": "Unprocessable entity.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"description": "Self registration for new users.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/NewRegisteredUser"
							}
						}
					}
				}
			}
		},
		"/v1/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Success.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithTokens"
								}
							}
						}
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Incorrect email or password.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"204": {
						"description": "Success."
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/refresh-tokens": {
			"post": {
				"operationId": "RefreshTokens",
				"responses": {
					"200": {
						"description": "Success.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserWithTokens"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"required": [
									"refreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/forgot-password": {
			"post": {
				"operationId": "ForgotPassword",
				"responses": {
					"204": {
						"description": "Success."
					},
					"404": {
						"description": "Not Found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"204": {
						"description": "Success."
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									}
								},
								"required": [
									"password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/send-verification-email": {
			"post": {
				"operationId": "SendVerificationEmail",
				"responses": {
					"204": {
						"description": "Success."
					},
					"401": {
						"description": "Unauthorized.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/v1/auth/verify-email": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"204": {
						"description": "Success."
					},
					"400": {
						"description": "Validation failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"401": {
						"description": "Email verification failed.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					},
					"404": {
						"description": "Not Found.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiError"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}